+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Added endpoints for user agents pause/unpause:
 
+  * GET `/users/me/agents`
   * POST `/users/me/agents/pause`
   * POST `/users/me/agents/unpause`